# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SerialCuts"
  },
  "paths": {
...
    "/api/seasons/import": {
      "post": {
        "summary": "Import Season Endpoint",
        "operationId": "import_season_endpoint_api_seasons_import_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SeasonImportRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImportResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/seasons/{season_id}": {
      "delete": {
        "summary": "Delete Season Endpoint",
        "operationId": "delete_season_endpoint_api_seasons__season_id__delete",
        "parameters": [
          {
            "name": "season_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Season Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/episodes/{episode_id}/story-context": {
      "get": {
        "summary": "Read Story Context",
        "operationId": "read_story_context_api_episodes__episode_id__story_context_get",
        "parameters": [
          {
            "name": "episode_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Episode Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StoryContextRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "summary": "Update Story Context",
        "operationId": "update_story_context_api_episodes__episode_id__story_context_put",
        "parameters": [
          {
            "name": "episode_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Episode Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StoryContextUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StoryContextRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/episodes/{episode_id}/outline": {
      "get": {
        "summary": "Read Episode Outline",
        "operationId": "read_episode_outline_api_episodes__episode_id__outline_get",
        "parameters": [
          {
            "name": "episode_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Episode Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EpisodeOutlineRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/jobs/{job_id}/cancel": {
      "post": {
        "tags": [
          "queue"
        ],
        "summary": "Cancel Job",
        "operationId": "cancel_job_api_jobs__job_id__cancel_post",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/jobs/{job_id}/retry": {
      "post": {
        "tags": [
          "queue"
        ],
        "summary": "Retry Job Endpoint",
        "operationId": "retry_job_endpoint_api_jobs__job_id__retry_post",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/jobs/{job_id}/retry-stage": {
      "post": {
        "tags": [
          "queue"
        ],
        "summary": "Retry Job From Stage Endpoint",
        "operationId": "retry_job_from_stage_endpoint_api_jobs__job_id__retry_stage_post",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Job Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JobStageRetryRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/jobs/{job_id}": {
      "delete": {
        "tags": [
          "queue"
        ],
        "summary": "Delete Job Endpoint",
        "operationId": "delete_job_endpoint_api_jobs__job_id__delete",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/jobs/{job_id}/stages": {
      "get": {
        "tags": [
          "queue"
        ],
        "summary": "Job Stages",
        "operationId": "job_stages_api_jobs__job_id__stages_get",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/JobStageRead"
                  },
                  "title": "Response Job Stages Api Jobs  Job Id  Stages Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "EpisodeOutlineRead": {
        "properties": {
          "episode_id": {
            "type": "integer",
            "title": "Episode Id"
          },
          "summary_json": {
            "additionalProperties": true,
            "type": "object",
            "title": "Summary Json"
          }
        },
        "type": "object",
        "required": [
          "episode_id",
          "summary_json"
        ],
        "title": "EpisodeOutlineRead"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ImportFileErrorRead": {
        "properties": {
          "file_name": {
            "type": "string",
            "title": "File Name"
          },
          "reason": {
            "type": "string",
            "title": "Reason"
          }
        },
        "type": "object",
        "required": [
          "file_name",
          "reason"
        ],
        "title": "ImportFileErrorRead"
      },
      "ImportResponse": {
        "properties": {
          "season_id": {
            "type": "integer",
            "title": "Season Id"
          },
          "created": {
            "type": "integer",
            "title": "Created"
          },
          "skipped_duplicates": {
            "type": "integer",
            "title": "Skipped Duplicates"
          },
          "episode_ids": {
            "items": {
              "type": "integer"
            },
            "type": "array",
            "title": "Episode Ids"
          },
          "scanned": {
            "type": "integer",
            "title": "Scanned",
            "default": 0
          },
          "errors": {
            "items": {
              "$ref": "#/components/schemas/ImportFileErrorRead"
            },
            "type": "array",
            "title": "Errors"
          }
        },
        "type": "object",
        "required": [
          "season_id",
          "created",
          "skipped_duplicates",
          "episode_ids"
        ],
        "title": "ImportResponse"
      },
      "JobRead": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "episode_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Episode Id"
          },
          "kind": {
            "type": "string",
            "title": "Kind"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "current_stage": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current Stage"
          },
          "progress": {
            "type": "number",
            "title": "Progress"
          },
          "progress_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Progress Message"
          },
          "error_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Message"
          },
          "started_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "finished_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "episode_id",
          "kind",
          "status",
          "current_stage",
          "progress",
          "error_message",
          "created_at",
          "updated_at"
        ],
        "title": "JobRead"
      },
      "JobStageRead": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "job_id": {
            "type": "integer",
            "title": "Job Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "started_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "finished_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          },
          "error_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Message"
          },
          "artifact_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Artifact Path"
          }
        },
        "type": "object",
        "required": [
          "id",
          "job_id",
          "name",
          "status",
          "started_at",
          "finished_at",
          "error_message",
          "artifact_path"
        ],
        "title": "JobStageRead"
      },
      "JobStageRetryRequest": {
        "properties": {
          "stage_name": {
            "type": "string",
            "pattern": "^(stage2_media|stage3_candidates|auto_export|render_clip|render_story_arc)$",
            "title": "Stage Name"
          }
        },
        "type": "object",
        "required": [
          "stage_name"
        ],
        "title": "JobStageRetryRequest"
      },
...
      "SeasonImportRequest": {
        "properties": {
          "root_path": {
            "type": "string",
            "minLength": 1,
            "title": "Root Path"
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          }
        },
        "type": "object",
        "required": [
          "root_path"
        ],
        "title": "SeasonImportRequest"
      },
...
      "StoryContextRead": {
        "properties": {
          "season_id": {
            "type": "integer",
            "title": "Season Id"
          },
          "episode_id": {
            "type": "integer",
            "title": "Episode Id"
          },
          "season_context": {
            "type": "string",
            "title": "Season Context"
          },
          "episode_summary": {
            "type": "string",
            "title": "Episode Summary"
          },
          "required_events": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Required Events"
          },
          "excluded_events": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Excluded Events"
          },
          "spoilers_allowed": {
            "type": "boolean",
            "title": "Spoilers Allowed"
          },
          "candidate_mode": {
            "type": "string",
            "title": "Candidate Mode"
          }
        },
        "type": "object",
        "required": [
          "season_id",
          "episode_id",
          "season_context",
          "episode_summary",
          "required_events",
          "excluded_events",
          "spoilers_allowed",
          "candidate_mode"
        ],
        "title": "StoryContextRead"
      },
      "StoryContextUpdate": {
        "properties": {
          "season_context": {
            "type": "string",
            "maxLength": 8000,
            "title": "Season Context",
            "default": ""
          },
          "episode_summary": {
            "type": "string",
            "maxLength": 8000,
            "title": "Episode Summary",
            "default": ""
          },
          "required_events": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 30,
            "title": "Required Events"
          },
          "excluded_events": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 30,
            "title": "Excluded Events"
          },
          "spoilers_allowed": {
            "type": "boolean",
            "title": "Spoilers Allowed",
            "default": true
          },
          "candidate_mode": {
            "type": "string",
            "pattern": "^(highlights|story)$",
            "title": "Candidate Mode",
            "default": "highlights"
          }
        },
        "type": "object",
        "title": "StoryContextUpdate"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}